pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "py-cord"
authors = [
    {name = "Pycord Development"}
]
description = "A Python wrapper for the Discord API"
readme = {content-type = "text/x-rst", file = "README.rst"}
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
#    "License-Expression :: MIT",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
    "aiohttp>=3.6.0,<4.0",
    "colorlog~=6.9.0",
    "typing-extensions>=4,<5",
]

[project.urls]
Homepage = "https://pycord.dev"
Changelog = "https://docs.pycord.dev/en/master/changelog.html"
Source = "https://github.com/Pycord-Development/pycord"
Documentation = "https://docs.pycord.dev"
Tracker = "https://github.com/Pycord-Development/pycord/issues"
Funding = "https://patreon.com/pycord"

[project.optional-dependencies]
voice = [
    "pynacl>=1.3.0,<1.6",
]
speed = [
    "aiohttp[speedups]>=3.11.18",
    "msgspec~=0.19.0",
]

[dependency-groups]
docs = [
    "furo==2024.8.6",
    "levenshtein==0.27.1",
    "myst-parser[linkify]==4.0.1",
    "sphinx==8.0.2",
    "sphinx-autodoc-typehints==2.5.0",
    "sphinx-copybutton==0.5.2",
    "sphinx-intl==2.3.2",
    "sphinxcontrib-trio==1.1.2",
    "sphinxcontrib-websupport==2.0.0",
    "sphinxext-opengraph==0.13.0",
    "typing-extensions==4.15.0",
]
dev = [
    "codespell==2.4.1",
    "coverage~=7.8",
    "mypy~=1.17.1",
    "pre-commit==4.3.0",
    "pytest~=8.4.1",
    "pytest-asyncio~=0.26.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.11.9",
]
ci = [
    "pygithub>=2.7.0",
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "discord/_version.py"

[tool.hatch.build.targets.wheel]
only-packages = true
packages = ["discord"]
only-include = ["discord", "README.rst", "LICENSE", "pyproject.toml", "discord/bin/*.dll", "discord/emojis.json", "discord/ibanner.txt", "discord/banner.txt", "discord/chan/ART.txt",]

[tool.hatch.build.targets.sdist]
only-packages = true
packages = ["discord"]
only-include = ["discord", "README.rst", "LICENSE", "pyproject.toml", "discord/bin/*.dll", "discord/emojis.json", "discord/ibanner.txt", "discord/banner.txt", "discord/chan/ART.txt",]

[tool.ruff]
target-version = "py310"
line-length = 120
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
exclude = [
    "discord/_version.py"
]

[tool.ruff.lint]
select = ["ALL"]
per-file-ignores = {}
# When ignoring a rule globally, please consider if it can be ignored in a more specific way.
# Also, leave a comment explaining why the rule is ignored.
extend-ignore = [
    "A001", # Ignored when migrating from original codebase, idem for others if not mentioned otherwise
    "PLW1641", # Object does not implement `__hash__` method
    "A002",
    "A005",
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN201",
    "ANN202",
    "ANN204",
    "ANN205",
    "ANN206",
    "ANN401",
    "ARG001",
    "ARG002",
    "ARG003",
    "ARG004",
    "ARG005",
    "ASYNC230",
    "BLE001",
    "C416",
    "C417",
    "C420",
    "C901",
    "COM812",
    "D100",
    "D101",
    "D102",
    "D103",
    "D105",
    "D107",
    "D202",
    "D205",
    "D212",
    "D214",
    "D215",
    "D400",
    "D401",
    "D402",
    "D404",
    "D409",
    "D412",
    "D413",
    "D415",
    "D416",
    "D417",
    "DTZ003",
    "DTZ004",
    "DTZ005",
    "E501",
    "E712",
    "E713",
    "E721",
    "E731",
    "E741",
    "EM101",
    "EM102",
    "EM103",
    "ERA001",
    "F401",
    "F402",
    "F403",
    "F405",
    "FBT001",
    "FBT002",
    "FBT003",
    "FIX002",
    "FURB167",
    "FURB188",
    "ISC001",
    "N801",
    "N803",
    "N804",
    "N806",
    "N811",
    "N818",
    "PERF102",
    "PERF203",
    "PERF401",
    "PGH003",
    "PLC0105",
    "PLR0206",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR1704",
    "PLR1714",
    "PLR2004",
    "PLR5501",
    "PLW0602",
    "PLW0603",
    "PLW2901",
    "PTH100",
    "PTH107",
    "PTH110",
    "PTH118",
    "PTH120",
    "PTH122",
    "PTH123",
    "PYI016",
    "PYI019",
    "PYI024",
    "PYI032",
    "PYI034",
    "PYI036",
    "PYI041",
    "PYI046",
    "RET501",
    "RET502",
    "RET503",
    "RET504",
    "RET505",
    "RET506",
    "RET507",
    "RSE102",
    "RUF001",
    "RUF005",
    "RUF006",
    "RUF012",
    "RUF013",
    "RUF015",
    "RUF021",
    "RUF022",
    "RUF023",
    "RUF046",
    "RUF051",
    "S101",
    "S110",
    "S307",
    "S311",
    "S603",
    "SIM102",
    "SIM105",
    "SIM108",
    "SIM114",
    "SIM115",
    "SIM118",
    "SIM201",
    "SIM212",
    "SIM910",
    "T201",
    "TC001",
    "TC003",
    "TC004",
    "TC006",
    "TD002",
    "TD003",
    "TD006",
    "TID252",
    "TRY002",
    "TRY003",
    "TRY201",
    "TRY300",
    "TRY301",
    "UP006",
    "UP031",
    "UP032",
    "UP035",
    "A004",
    "D104",
    "D200",
    "D213",
    "D406",
    "F541",
    "F811",
    "FA100",
    "INP001",
    "S104",
    "SIM116",
    "SLF001",
    "W293",
    "D211",
    "D203" # conflicts with formatter
]

[tool.mypy]
namespace_packages = true
install_types = true
strict = true
show_error_codes = true
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
